Edit for puppet-lint and add onlyif

+++ 0x0F-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,6 @@
 # Kills a process named killmenow
 exec { 'killmenow':
   command => 'pkill -f killmenow',
-  path => '/usr/bin:/usr/sbin:/bin'
+  path    => '/usr/bin:/usr/sbin:/bin',
+  onlyif  => 'pgrep -f killmenow'
 }
